GH-46691: [CI][Packaging] Update platform tag on generated wheel name to match newest auditwheel naming (#46705)
### Rationale for this change

The new version of auditwheel has added improvements to detect libc / platform on the wheels:
- https://github.com/pypa/auditwheel/pull/548

This has updated the ordering of the platform tags for some of the generated wheels. For the case of our manylinux_2014 and libc 2.17 but only for Python 3.13 amd64 and 3.13t arm64. The rest are using the old order.

### What changes are included in this PR?

Force the newest version and update to new order of platform tags.

### Are these changes tested?

Via archery.

### Are there any user-facing changes?

No

* GitHub Issue: #46691

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>