GH-46636: [R] Fix evaluation of external objects not in global environment in `case_when()` (#46667)
### Rationale for this change

When a script is called in an environment that isn't the global environment (for instance with `source("my-script.R", local = new.env())`, `case_when()` would fail to detect external objects used in conditions.

This PR fixes this behavior.

Fixes #46636 

### What changes are included in this PR?

When evaluating expressions in `dplyr` functions, `eval_tidy()` now takes into account `mask` as an environment where it should look for external objects.

@ thisisnic suggested in #46636 that the bug might be due to https://github.com/apache/arrow/blob/main/r/R/dplyr-funcs-conditional.R#L116 but I couldn't find a way to fix it there.

### Are these changes tested?

I added a test for this scenario. I ensured it failed before the change and succeeds after.

### Are there any user-facing changes?

There is one user-facing, non-breaking change, illustrated both in the related issue and in the new test.

* GitHub Issue: #46636

Authored-by: etiennebacher <etienne.bacher@protonmail.com>
Signed-off-by: Nic Crane <thisisnic@gmail.com>