GH-46704: [C++] Fix OSS-Fuzz build failure (#46706)
### Rationale for this change

PR #46408 included a typo that changed list-view IPC tests to use the same data as list tests. This was detected as a duplicate corpus file by the OSS-Fuzz CI build.

### What changes are included in this PR?

Undo mistake that led to using the same test data for lists and list-views. Also fix a regression in the CUDA tests, due to reading non-CPU memory when fetching the first offset in a list/binary array.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

No.

* GitHub Issue: #46704

Authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>