GH-46827: [C++] Update Meson Configuration for compute shared lib (#46839)
### Rationale for this change

The major refactor of the compute sources in https://github.com/apache/arrow/pull/46261 addressed updates to the CMake configuration but not Meson. This is a follow up to get the Meson builds working again

### What changes are included in this PR?

Meson configuration files are updated to reflect new source structure. gtest has also been bumped to a new WrapDB version, which fixes some undefined behavior that was compounded by updates to the compute test structure

### Are these changes tested?

Yes 

### Are there any user-facing changes?

No
* GitHub Issue: #46827

Authored-by: Will Ayd <william.ayd@icloud.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>