GH-40754: [Python] Expose tls_ca_file_path to S3FileSystem (#45881)
### Rationale for this change

Currently, when using the pyarrow.fs.S3FileSystem, it is not immediately obvious how provide an alternative TLS certificate authority when working with (for example) a self-hosted minio using a self-signed certificate. It is possible to do this via environment variables, as per #40754; however, the user may not be able to easily specify these depending on their python execution environment.

### What changes are included in this PR?

Adds an additional parameter to S3FileSystem to specify the file path to a custom certificate.

### Are these changes tested?

Yes. A unit-test for S3FileSystem pickling is included; and I have manually verified that I can access files from a self-hosted minio with TLS enabled.

### Are there any user-facing changes?

An additional parameter `tls_ca_file_path` is available on the S3FileSystem constructor.

* GitHub Issue: #40754

Authored-by: bw513 <192128226+bw513@users.noreply.github.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>