GH-46833: [Python] Expose ConfigureManagedIdentityCredential and ConfigureClientSecretCredential to AzureFileSystem on PyArrow (#46837)
### Rationale for this change

`ClientSecretCredential` is another method to authenticate to Azure resources using Service Principle (SP). We work on AzureML, where there are two main types of compute ComputeInstance (personal VM) and ComputeCluster. In context of AzureML you can work interactively on ComputeInstance, where pyarrow  defaults to`DefaultAzureCredential` when reading/writing data over `abfss://` protocol. However when running AzureML "jobs" which is run non-interactively execution, the context (network) is different and `DefaultAzureCredential` doesn't work. An alternative method for non-interactive job is to use Service Principle i.e `ClientSecretCredential`.  This way we can create `AzureFileSystem` object and give it pyarrow `dataset`

### What changes are included in this PR?

Adding ConfigureClientSecretCredential to AzureFilesystem, which is already implemented at the C++, but hasn't been propagated to python library. This pull request just propagates C++ method to pyarrow

### Are these changes tested?

Yes, test have been included in the relevant file 

### Are there any user-facing changes?

Docs of `AzureFileSystem` have been updated in the code. It would be amazing for those changes to surface on the website, which I'm happy to try to help with if needed

* GitHub Issue: #46833

Lead-authored-by: Kirill Tsyganov <kirill.tsyganov@outlook.com.au>
Co-authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>