GH-35166: [C++][Compute] Increase precision of decimals in sum aggregates (#44184)
### Rationale for this change

As documented in #35166, when Arrow performs a sum, product, or mean of an array of type `decimalX(P, S)`, it returns a scalar of type `decimalX(P, S)`. This is true even if the aggregate does not fit in the specified precision. For instance, a sum of two `decimal128(1, 0)`'s such as `1 + 9` is a `decimal128(2, 0)`. But (in Python):

```python
import pyarrow as pa
from decimal import Decimal

arr = pa.array([Decimal("1"), Decimal("9")], type=pa.decimal128(1, 0))
assert arr.sum().type == pa.decimal128(1, 0)
```

This is recognized in the rules for _binary_ addition and multiplication of decimals (see footnote 1 in [this section](https://arrow.apache.org/docs/cpp/compute.html#arithmetic-functions)), but this does not apply to array aggregates.

In #35166 I did a bit of research following a question from @ westonpace , and it seems that there's no standard approach to this across DBMS's, but a common solution is to set the precision of the result of a sum to the maximum possible precision of the underlying type. That is, a sum of `decimal128(1, 0)`'s becomes a `decimal128(38, 0)`.

However, products and means differ further. For instance, in both instances, duckdb converts a decimal to a double, which makes sense as the precision of the product of an array of decimals would likely be huge, e.g., an array of size N with precision 2 decimals would have precision at least 2^N.

This PR implements the minimum possible change: replace the return types of the sum aggregate of `decimal128(P, S)` to `decimal128(38, S)`, `decimal256(P, S)` to `decimal256(76, S)`, `decimal32(P, S)` to `decimal32(9, S)`, and `decimal64(P, S)` to `decimal64(18, S)`.

### What changes are included in this PR?

- [x] Update C++ kernels to support the change
- [x] Update docs to reflect the change
- [x] Fix tests in the languages that depend on the C++ engine
- [x] Determine if there are other languages which do _not_ depend on the C++ engine which should also be updated

### Are these changes tested?

They are tested in the following implementations:

- [x] Python
- [x] C++

### Are there any user-facing changes?

Yes. This changes the return type of a scalar aggregate of decimals.

**This PR includes breaking changes to public APIs.**

Specifically, the return type of a scalar aggregate of decimals changes. This is unlikely to break downstream applications as the underlying data has not changed, but if an application relies on the (incorrect!) type information for some reason, it would break.
* GitHub Issue: #35166

Lead-authored-by: Kevin H Wilson <khwilson@gmail.com>
Co-authored-by: Rossi Sun <zanmato1984@gmail.com>
Co-authored-by: Kevin Wilson <khwilson@gmail.com>
Co-authored-by: Antoine Pitrou <pitrou@free.fr>
Signed-off-by: Rossi Sun <zanmato1984@gmail.com>