GH-46777: [C++] Use SimplifyIsIn only when the value_set of the expression is lower than a threshold (#46859)
### Rationale for this change

Using `SimplifyIsIn` when the value set is large has a substantial performance penalty.

### What changes are included in this PR?

Ensure we do not use the simplification when the value_set on the expression is higher than a threshold (50).

### Are these changes tested?

I've tested locally that the reproducer goes back to pre change levels.
```
$ python read.py 
=== PYARROW VERSION 20 ===
Retrieved 10,000,000 rows in 3.08 seconds.
```

I have added a test for large sets and validate the expression is not being modified.

### Are there any user-facing changes?

No

* GitHub Issue: #46777

Lead-authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Co-authored-by: Antoine Pitrou <pitrou@free.fr>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>