GH-46272: [C++] Build Arrow libraries with `-Wmissing-definitions` on gcc
The warning option `-Wmissing-declarations` allows finding private functions that erroneously have global linkage
(because they are neither static nor in the anonymous namespace).

We only care about this for the public Arrow libraries, not for tests or utilities where it's harmless to have
private functions that nevertheless have global linkage (and changing that would require a lot of pointless
code churn).