GH-47074: [Release] Use reproducible mtime for csharp/ in source archive (#47076)
### Rationale for this change

The current source archive creation is reproducible when we use the same Git working tree.

But it's not reproducible when we use different Git working trees.

### What changes are included in this PR?

Use the committer date of the target commit instead of the `charp/` mtime in the current Git working tree for `csharp/` in the source archive.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

No.
* GitHub Issue: #47074

Authored-by: Sutou Kouhei <kou@clear-code.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>