Numba interop: fix device pointer value with NVIDIA bindings
Numba and Numba-CUDA return a different type from `Context.memalloc()`
depending on whether their built-in ctypes bindings or the NVIDIA CUDA
Python bindings are in use - either e `ctypes.c_void_p` or a
`cuda.bindings.driver.CUdeviceptr`. Whilst this inconsistency is
unfortunate, it's hard to change as existing code in the wild depends on
it.

The issue in Arrow is that the value of the pointer cannot be obtained
by `device_pointer.value` for a `CUdeviceptr`. Numba and Numba-CUDA do
provide another property, `device_pointer_value`, that provides the
device pointer as an `int` regardless of the kind of binding in use, so
we can switch to use this for consistency between the two kinds of
bindings.

Fixes #47128.