MINOR: [Release] Make post-08-remove-old-artifacts.sh crossplatform (#47250)
### Rationale for this change

`dev/release/post-08-remove-old-artifacts.sh` uses [`tac`](https://man7.org/linux/man-pages/man1/tac.1.html) which may not be available on non-Linuxes. We should be able to get its same behavior with more common utilities.

### What changes are included in this PR?

Swaps the call to `tac` with `sort --version-sort` which should be equivalent.

### Are these changes tested?

Yes. This shouldn't impact which artifacts are removed, just the order in which they are.

### Are there any user-facing changes?

No.

Authored-by: Bryce Mecum <petridish@gmail.com>
Signed-off-by: Bryce Mecum <petridish@gmail.com>