GH-47371: [Python] Fix installation of Numba-CUDA in CI environment
As per the Numba-CUDA docs, `numba-cuda` should be installed with a
variant depending on the CUDA version. Failing to specify the variant
leads to `numba-cuda` being installed without the CUDA toolkit wheels
that it needs. This commit adds the variant specifier to the
installation.

I think Arrow is using CUDA 11.7 in CI images, so I've changed the
Numba-CUDA version to 0.18.1, which is the last version to support CUDA
11 (the Numba-CUDA policy is to support all minor versions within the
last two major versions of CUDA). IF Arrow moves to CUDA 12 for CI
images, this could be changed back to `latest` in future.

Numba-CUDA installation docs: https://nvidia.github.io/numba-cuda/user/installation.html