Fix host handling for default HDFS URI (fix for #25324 when used via `from_uri()`
In #25324 a fix is introduced for the python HadoopFileSystem, but it does not work if you use `from_uri()`, as it is passed to the underlying C++ implementation of the options parsing. The "default" case is not handled as in the python case, as the whole "hdfs://default" is passed to the underlying hdfs library, that expect "default" to search in $HADOOP_CONF_DIR/core-site.xml.