GH-47179: [Python] Revert FileSystem.from_uri to be a staticmethod again (#47178)
### Rationale for this change

It seems as part of #45089 `FileSystem.from_uri` was changed from a static function to a regular method. I believe this to be an error unintentionally introduced as part of the PR, based on the fact that this was not mentioned to be an explicit change, and the documentation attached to the function still shows it being called as a static method.

This changed was released as part of version 21.0. Previously, it was a static method on the class.

Now, when attempting to call `FileSystem.from_uri`, it breaks in two different cases:
1. Calling it as a class static method with an explicit `uri` argument.
2. Calling it as an instance method.

However, calling it as a class static method without an explicit `uri` argument (ex. `FileSystem.from_uri("/abc")`) still works.

This PR adds the `@ staticmethod` annotation back to this method.

### What changes are included in this PR?

`FileSystem.from_uri` is now a static method again.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

Yes, to the `FileSystem.from_uri` function.

* GitHub Issue: #47179

Authored-by: ff-kamal <kamal@featureform.com>
Signed-off-by: AlenkaF <frim.alenka@gmail.com>