GH-47485: [C++][CI] Work around Valgrind failure on Azure tests
The Azure tests fail because of a likely bug in the Azure SDK, which tries to deallocate a curl shared handler while it's still in use, and doesn't report deallocation errors (reported upstream at https://github.com/Azure/azure-sdk-for-cpp/issues/6722).

This PR adds a dedicated Valgrind suppression to ignore the corresponding memory leak.