GH-47485: [C++][CI] Work around Valgrind failure on Azure tests (#47496)
### Rationale for this change

The Azure tests fail because of a likely bug in the Azure SDK, which tries to deallocate a curl shared handler while it's still in use, and doesn't report deallocation errors (reported upstream at https://github.com/Azure/azure-sdk-for-cpp/issues/6722).

### What changes are included in this PR?

Add a dedicated Valgrind suppression to ignore the corresponding memory leak.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

No.
* GitHub Issue: #47485

Authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>