GH-46669: [CI][Archery] Automate Zulip and email notifications for Extra CI (#47546)
### Rationale for this change

We plan to move more Crossbow CI jobs to the main Arrow repository. We have some reporting tools like the Zulip message notifications, the email report notifications and the nightlies report page: http://crossbow.voltrondata.com/ that we should replicate for the CI.

### What changes are included in this PR?

Add GH action that will send the chat and email reports. Also it adds infrastructure to archery to work with Arrow CI.

There's a lot of baggage around crossbow so I've decided to implement a new module for Arrow's `ci` where we can use the GitHub API and we don't require cloning and a lot of things that were required to work with crossbow. In the future we might be able to get rid of all the related `crossbow` CLI part.

### Are these changes tested?

Via CI

### Are there any user-facing changes?

No

* GitHub Issue: #46669

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>