GH-47560: [C++] Fix host handling for default HDFS URI (#47458)
### Rationale for this change

In #25324 a fix is introduced for the python HadoopFileSystem, but it does not work if you use `from_uri()`, as it is passed to the underlying C++ implementation of the options parsing. The "default" case is not handled as in the python case, as the whole "hdfs://default" is passed to the underlying hdfs library, that expect "default" to search in `$HADOOP_CONF_DIR/core-site.xml`.

### What changes are included in this PR?

Handle the `HadoopFileSystem.from_uri()` (or `FileSystem.from_uri()` when using `hdfs://default:xxx`) special HDFS URIs.

### Are these changes tested?

There are no specific tests for this feature, but existing HDFS CI jobs pass.

### Are there any user-facing changes?

Not exactly, but the documentation is honored for the `from_uri()` case.

* GitHub Issue: #47560

Lead-authored-by: Diego Sevilla Ruiz <dsevilla@um.es>
Co-authored-by: Antoine Pitrou <pitrou@free.fr>
Signed-off-by: Antoine Pitrou <antoine@python.org>