[C++] Patch vendored pcg library to enable msvc arm64 intrinsics
This enables building Arrow C++ for Windows ARM64 with MSVC when setting
ARROW_SIMD_LEVEL to NONE (to disable xsimd usage, which does not support
msvc arm64 intrinsics, and is non-trivial to fix).

This patch is based on https://github.com/imneme/pcg-cpp/pull/99.
The upstream pcg library has not been updated in the past 3 years, so
this may never get merged.